requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
Werkzeug==2.2.3
email-validator==2.0.0
scrapy==2.9.0
selenium==4.11.2
webdriver-manager==3.8.6
APScheduler==3.10.1
python-dotenv==1.0.0